Moved location of DashboardMappings.db.
